feat: adicionado store para estId e estUrl
