+++ package.json
@@ -26,7 +26,7 @@
     "tailwindcss-animate": "^1.0.7",
     "validator": "^13.11.0",
     "web-vitals": "^2.1.4",
-    "zustand": "^4.4.4"
+    "zustand": "^4.5.1"
   },
   "scripts": {
     "dev": "react-scripts start",
