+++ src/App.css
@@ -17,6 +17,7 @@
 html, body, #root {
   height: 100%;
 }
+
 /* 
 section{
   padding: 0 10%;
@@ -30,9 +31,10 @@ section{
 
 @layer base {
   :root { 
-    --background: 0 0% 100%;
-  /*   --background: 224 71% 4%; */
+/*     --background: 0 0% 100%;
     --foreground: 222.2 47.4% 11.2%;
+    --background: 224 71% 4%;
+    --foreground: 213 31% 91%;
 
     --text: 0, 0%, 100%;
  
@@ -47,7 +49,7 @@ section{
  
     --card: 0 0% 100%;
     --card-foreground: 222.2 47.4% 11.2%;
-   /*  --primary: #3498db; */
+   
 
     --primary: 204, 70%, 53%;
     --primary-foreground: 210 40% 98%;
@@ -62,6 +64,36 @@ section{
     --destructive-foreground: 210 40% 98%;
  
     --ring: 215 20.2% 65.1%;
+ 
+    --radius: 0.5rem; */
+    --background: 224 71% 4%;
+    --foreground: 213 31% 91%;
+ 
+    --muted: 223 47% 11%;
+    --muted-foreground: 215.4 16.3% 56.9%;
+ 
+    --accent: 216 34% 17%;
+    --accent-foreground: 210 40% 98%;
+ 
+    --popover: 224 71% 4%;
+    --popover-foreground: 215 20.2% 65.1%;
+ 
+    --border: 216 34% 17%;
+    --input: 216 34% 17%;
+ 
+    --card: 224 71% 4%;
+    --card-foreground: 213 31% 91%;
+ 
+    --primary: 204, 70%, 53%;
+    --primary-foreground: 222.2 47.4% 1.2%;
+ 
+    --secondary: 222.2 47.4% 11.2%;
+    --secondary-foreground: 210 40% 98%;
+ 
+    --destructive: 0 63% 31%;
+    --destructive-foreground: 210 40% 98%;
+ 
+    --ring: 216 34% 17%;
  
     --radius: 0.5rem;
   }
@@ -99,7 +131,7 @@ section{
     --radius: 0.5rem;
   }
 }
- 
+ /*  --primary: #3498db; */
 @layer base {
   * {
     @apply border-border;
