+++ src/index.js
@@ -9,7 +9,7 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode >
     <QueryClientProvider client={queryClient}>
-      <App />
+      <App/>
     </QueryClientProvider>
   </React.StrictMode>
 );
