+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "allowJs": true,
     "checkJs": false,
-    "target": "es5",
+    "target": "ES6",
     "module": "commonjs",
     "strict": false,
     "esModuleInterop": true,
@@ -13,6 +13,6 @@
       "@/*": ["*"]
     }
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "./stores/**/*",  "stores/**/*"],
   "exclude": ["node_modules"]
 }
